myFunction.py
You will need to create four functions:
● HotelCost - This function will take the number of nights a user will be
staying at a hotel as an argument and return a total cost for the hotel stay
(You can choose the price per night charged at the hotel).
● PlaneCost - This function will take the city you are flying to as an argument
and return a cost for the flight (Hint: use if/else if statements in the function
to retrieve a price based on the chosen city).
● CarRental - This function will take the number of days the car will be hired
for as an argument and return the total cost of the car rental.
● HolidayCost - This function will take three arguments: the number of nights
a user will be staying in a hotel, the city the user will be flying to and the
number of days that the user will be hiring a car for. Using these three
arguments, you can call all three of the above functions with respective
arguments and finally return a total cost for your holiday.
● Print out the value of your Holiday function to see the result!
● Try using your app with different combinations of input to show it’s
compatibility with different options.